+++ CHANGES.rst
@@ -6,6 +6,9 @@ Changelog
 
 - Added `security.check_zope_admin` (moved from `Products.CPUtils`).
   [gbastien]
+- Added `workflow.update_role_mappings_for` helper to update WF role mappings
+  for a given object.
+  [gbastien]
 
 0.69 (2023-05-31)
 -----------------